requirements.txt
pytest==8.3.3
requests==2.32.3
allure-pytest==2.13.5
allure-python-commons==2.13.5
